Copy edits from MaxLAI branch (5/19/21, 5/20/21 commits)

+++ src/ISiteCohorts.cs
@@ -15,7 +15,7 @@ namespace Landis.Library.PnETCohorts
 
         float[] FolResp { get; }
         
-        byte CanopyLAImax{get;}
+        float CanopyLAImax{get;}
 
         int AverageAge { get; }
